Refactor out generically useful bit of fetch.h into curl.h
