+++ src/http/curl.h
@@ -1,0 +1,70 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the Apache 2.0 License.
+#pragma once
+
+#include "ccf/ds/logger.h"
+#include "ccf/ds/nonstd.h"
+
+#include <curl/curl.h>
+#include <memory>
+#include <span>
+
+#define CHECK_CURL_EASY(fn, ...) \
+  do \
+  { \
+    const auto res = fn(__VA_ARGS__); \
+    if (res != CURLE_OK) \
+    { \
+      throw std::runtime_error(fmt::format( \
+        "Error calling " #fn ": {} ({})", res, curl_easy_strerror(res))); \
+    } \
+  } while (0)
+
+#define CHECK_CURL_EASY_SETOPT(handle, info, arg) \
+  CHECK_CURL_EASY(curl_easy_setopt, handle, info, arg)
+#define CHECK_CURL_EASY_GETINFO(handle, info, arg) \
+  CHECK_CURL_EASY(curl_easy_getinfo, handle, info, arg)
+
+namespace ccf::curl
+{
+
+  class UniqueCURL
+  {
+  protected:
+    std::unique_ptr<CURL, void (*)(CURL*)> p;
+
+  public:
+    UniqueCURL() : p(curl_easy_init(), [](auto x) { curl_easy_cleanup(x); })
+    {
+      if (!p.get())
+      {
+        throw std::runtime_error("Error initialising curl easy request");
+      }
+    }
+
+    operator CURL*() const
+    {
+      return p.get();
+    }
+  };
+
+  class UniqueSlist
+  {
+  protected:
+    std::unique_ptr<curl_slist, void (*)(curl_slist*)> p;
+
+  public:
+    UniqueSlist() : p(nullptr, [](auto x) { curl_slist_free_all(x); }) {}
+
+    void append(const char* str)
+    {
+      p.reset(curl_slist_append(p.release(), str));
+    }
+
+    curl_slist* get() const
+    {
+      return p.get();
+    }
+  };
+
+} // namespace ccf::curl
